Move reference test fixtures and fix linebreak normalization tests
