+++ test/fixtures/dictionaries/references/cyclic_references/input.json
@@ -1,5 +1,5 @@
 {
-	"name": "ARGO Dictionary",
+	"name": "Test Dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
@@ -15,7 +15,7 @@
 						"key": true
 					},
 					"restrictions": {
-						"regexp": "^[\\w\\s\\W]{5,}$"
+						"regex": "#/ID_REG_EXP"
 					}
 				},
 				{
@@ -23,7 +23,7 @@
 					"valueType": "string",
 					"description": "Donor Biological Sex",
 					"restrictions": {
-						"codeList": ["Male", "Female", "Unknown", "No answer", "Other"]
+						"codeList": ["#/SEX", "Other"]
 					}
 				},
 				{
@@ -37,5 +37,13 @@
 				}
 			]
 		}
-	]
+	],
+	"references": {
+		"ID_REG_EXP": "^[\\w\\s\\W]{5,}$",
+		"MALE": "Male",
+		"FEMALE": "Female",
+		"OTHER": "#/UNKNOWN",
+		"UNKNOWN": "#/OTHER",
+		"SEX": ["#/MALE", "#/FEMALE", "#/OTHER"]
+	}
 }
