+++ test/fixtures/dictionaries/references/empty_references_section/output.json
@@ -1,5 +1,5 @@
 {
-	"name": "ARGO Dictionary",
+	"name": "Test Dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
@@ -15,7 +15,7 @@
 						"key": true
 					},
 					"restrictions": {
-						"regexp": "#/ID_REG_EXP"
+						"script": "field.length > 5"
 					}
 				},
 				{
@@ -37,8 +37,5 @@
 				}
 			]
 		}
-	],
-	"references": {
-		"ID_REG_EXP": "^[\\w\\s\\W]{5,}$"
-	}
+	]
 }
