+++ test/fixtures/dictionaries/references/no_references_section/output.json
@@ -1,5 +1,5 @@
 {
-	"name": "ARGO Dictionary",
+	"name": "Test Dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
@@ -15,7 +15,7 @@
 						"key": true
 					},
 					"restrictions": {
-						"regexp": "^[\\w\\s\\W]{5,}$"
+						"script": "field.length > 5"
 					}
 				},
 				{
