+++ test/fixtures/dictionaries/references/non_existing_references/output.json
@@ -1,5 +1,5 @@
 {
-	"name": "ARGO Dictionary",
+	"name": "Test Dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
@@ -15,7 +15,7 @@
 						"key": true
 					},
 					"restrictions": {
-						"regexp": "#/ID_REG_EXP"
+						"regex": "^[\\w\\s\\W]{5,}$"
 					}
 				},
 				{
@@ -23,7 +23,7 @@
 					"valueType": "string",
 					"description": "Donor Biological Sex",
 					"restrictions": {
-						"codeList": ["#/SEX", "Other", "#/SELF_REFERENCE"]
+						"codeList": ["Male", "Female", "Other"]
 					}
 				},
 				{
@@ -37,12 +37,5 @@
 				}
 			]
 		}
-	],
-	"references": {
-		"ID_REG_EXP": "^[\\w\\s\\W]{5,}$",
-		"MALE": "Male",
-		"FEMALE": "Female",
-		"SELF_REFERENCE": "#/SELF_REFERENCE",
-		"SEX": ["#/MALE", "#/FEMALE"]
-	}
+	]
 }
