+++ test/fixtures/dictionaries/references/references_within_references/output.json
@@ -1,5 +1,5 @@
 {
-	"name": "ARGO Dictionary",
+	"name": "Test Dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
@@ -15,7 +15,7 @@
 						"key": true
 					},
 					"restrictions": {
-						"script": "field.length > 5"
+						"regex": "^[\\w\\s\\W]{5,}$"
 					}
 				},
 				{
@@ -23,7 +23,7 @@
 					"valueType": "string",
 					"description": "Donor Biological Sex",
 					"restrictions": {
-						"codeList": ["Male", "Female", "Other"]
+						"codeList": ["Male", "Female", "Unknown", "No answer", "Other"]
 					}
 				},
 				{
