+++ test/functional/schemaTest.ts
@@ -21,9 +21,9 @@ import { expect } from 'chai';
 import { normalizeSchema } from '../../src/services/schemaService';
 import input1 from '../fixtures/dictionaries/linebreak/input1';
 import input2 from '../fixtures/dictionaries/linebreak/input2';
-import output from '../fixtures/dictionaries/linebreak/input1';
+import output from '../fixtures/dictionaries/linebreak/output';
 
-describe('Verify different EOL are normalized', () => {
+describe('noarmalizeSchema', () => {
 	it('Should not alter already formatted files', () => {
 		const normalizedSchema = normalizeSchema(input1.schemas[0]);
 		expect(normalizedSchema).to.deep.eq(output.schemas[0]);
